requirements.txt
coverage==4.5.3
Django==2.2
pytz==2018.9
sqlparse==0.3.0
django-rest-swagger==2.2.0
coreapi==2.3.3
coreschema==0.0.4
jinja2==2.10
MarkupSafe==1.1.1
itypes==1.1.0
requests==2.21.0
certifi==2019.3.9
chardet==3.0.4
idna==2.8
urllib3==1.24.1
uritemplate==3.0.0
djangorestframework==3.9.2
openapi-codec==1.3.2
coreapi==2.3.3
coreschema==0.0.4
jinja2==2.10
MarkupSafe==1.1.1
itypes==1.1.0
requests==2.21.0
certifi==2019.3.9
idna==2.8
urllib3==1.24.1
uritemplate==3.0.0
simplejson==3.16.0
pip==19.0.3
psycopg2-binary==2.7.6.1
setuptools==40.8.0
wheel==0.33.1

